Refactor changelog entry to separate component
