+++ pages/index.tsx
@@ -7,6 +7,8 @@ import remarkGfm from 'remark-gfm'
 import remarkBreaks from 'remark-breaks'
 import rehypeRaw from 'rehype-raw'
 import emoji from 'remark-emoji';
+import {ChangelogEntry} from "../types/ChangelogEntry";
+import ChangelogCard from "../components/ChangelogCard";
 
 const Home: NextPage = () => {
   const [players, setPlayers] = useState(-1)
@@ -38,14 +40,14 @@ const Home: NextPage = () => {
     }
   }, [fetchPlayers])
 
-  const [changelogNews, setchangelogNews] = useState([])
+  const [changelogNews, setchangelogNews] = useState<ChangelogEntry[]>([])
 
   const fetchchangelogNews = useCallback(async () => {
-    const data = await axios.get(
+    const data = await axios.get<ChangelogEntry[]>(
       'https://changelog.qplay.cz/api/changelog'
     )
     if (data?.data) {
-      setchangelogNews(data.data.filter((item:any) => item.published))
+      setchangelogNews(data.data.filter((item) => item.published))
     }
   }, [])
 
@@ -53,48 +55,11 @@ const Home: NextPage = () => {
     fetchchangelogNews();
   }, []);
 
-  function getImage(text: string) {
-    const image = text.match(/src="(.*?)"/);
-    return image ? image[0].replace('src="', '').replace('"', '') : '/images/404image.png';
-  }
-
-  function replaceTags(text: string) {
-    return text.replace(/<br[^>]*>/gi, '').replace(/<p>(.*)<\/p>/g, '').replace(/<img .*?>/g, '');
-  }
-
-  function convertDate(date: string) {
-    return new Date(date).toLocaleString("cs-CZ", {
-      year: "numeric",
-      month: "long",
-      day: "numeric",
-      hour12: false
-    })
-  }
-
   function ChangeLogNewsRender() {
     const news = changelogNews.slice(0, 3).map(
       (element) => {
         return (
-          <div className="col-md-4 mb-4" key={element['id']}>
-            <div className="card">
-              <div className="card-image" style={{ backgroundImage: `url('` + getImage(element['notes']) + `')` }}></div>
-              <div className="card-body">
-                <span className="tag" style={{ backgroundColor: element['typecolor'] }}>{element['typename']}</span>
-                <h5 className="card-title">{element['headline']}</h5>
-                <p className="card-text">
-                  <ReactMarkdown remarkRehypeOptions={{allowDangerousHtml: true}} rehypePlugins={[rehypeRaw]} remarkPlugins={[remarkGfm, remarkBreaks, emoji]}>{replaceTags(element['notes'])}</ReactMarkdown>
-                </p>
-                <p className="mb-0 readmore">
-                  <Link href={`/seznam-zmen/` + element['id']} passHref>
-                    <a>Číst více</a>
-                  </Link>
-                </p>
-              </div>
-              <div className="card-footer">
-                {convertDate(element['timestamp'])}
-              </div>
-            </div>
-          </div>
+          <ChangelogCard key={element['id']} element={element} />
         )
       }
     )
