+++ types/ChangelogEntry.ts
@@ -1,0 +1,9 @@
+export interface ChangelogEntry {
+  id: number
+  notes: string
+  typecolor: string
+  typename: string
+  headline: string
+  timestamp: string
+  published: boolean
+}
